Many Router improvements, and consistent library... (#12)
- Moves much of the functionality that previously existed in the
FastRouteRouter instance into its own decorator object. This includes
all methods for shortcuts to HTTP request methods (get, post, put,
delete), the parsing of strings into ControllerAction objects, and the
mount() functionality. Passing a Router implementation into
FriendlyRouter will now allow these methods to be used on any Router
implementation, also improves the complexity in that class and makes it
more specific to integration with the FastRoute library.

- Adds a new resource() method on the FriendlyRouter that allows the
setting of all CRUD operations in one simple method call. Pass the name
of the resource (no inflection is done, it is used as-is) and the
controller class that should handle this resource. Actions and methods
map as follows:
  - GET /resource -> Controller#index
  - GET /resource/{id} -> Controller#show
  - POST /resource -> Controller#create
  - PUT /resource/{id} -> Controller#update
  - DELETE /resource/{id} -> Controller#delete

- Restructures code that integrates a library implementation to be more
consistent with its physical code structure.